fix: `TransactionScript.scriptDataLength` to `u32`

+++ packages/transactions/src/coders/transaction.ts
@@ -81,7 +81,7 @@ export class TransactionScriptCoder extends Coder<TransactionScript, Transaction
 
     parts.push(new U64Coder().encode(value.scriptGasLimit));
     parts.push(new NumberCoder('u32').encode(value.scriptLength));
-    parts.push(new NumberCoder('u16').encode(value.scriptDataLength));
+    parts.push(new NumberCoder('u32').encode(value.scriptDataLength));
     parts.push(new NumberCoder('u32').encode(value.policyTypes));
     parts.push(new NumberCoder('u8').encode(value.inputsCount));
     parts.push(new NumberCoder('u8').encode(value.outputsCount));
@@ -104,7 +104,7 @@ export class TransactionScriptCoder extends Coder<TransactionScript, Transaction
     const scriptGasLimit = decoded;
     [decoded, o] = new NumberCoder('u32').decode(data, o);
     const scriptLength = decoded;
-    [decoded, o] = new NumberCoder('u16').decode(data, o);
+    [decoded, o] = new NumberCoder('u32').decode(data, o);
     const scriptDataLength = decoded;
     [decoded, o] = new NumberCoder('u32').decode(data, o);
     const policyTypes = decoded;
